+++ internal/e2eselfhost/self_host_ssa_lift_diff_test.go
@@ -82,7 +82,7 @@ func TestSelfHostSSALiftDifferential(t *testing.T) {
 		{"strcat", `function main(): i32 { return ("foo" + "bar").len(); }`},
 		{"strbuf", `function main(): i32 { strbuf_reset(); strbuf_append("ab"); strbuf_append("cde"); return strbuf_take().len(); }`},
 		{"exitprog", `function main(): i32 { exit(42); return 0; }`},
-		{"strindex", `function main(): i32 { return "ABC"[1]; }`},
+		{"strindex", `function main(): i32 { return "ABC"[1] as i32; }`},
 		{"optval", `function main(): i32 { match (Some(42)) { Some(v) => { return v; }, None => { return 0; } } }`},
 		{"argslen", `function main(): i32 { return args().len(); }`},
 		{"closure", `function main(): i32 { var x = 10; var f = (y: i32) => x + y; return f(5); }`},
